requirements.txt
# Retail Insights Agent - Dependencies

# Google ADK (Multi-Agent Framework)
google-adk==1.15.1

# LLM Providers
openai==1.109.1

# SQL Generation (PandasAI + DuckDB)
pandasai==2.1
duckdb==0.10.3

# NVIDIA NeMo Guardrails
nemoguardrails==0.19.0
langchain==1.1.2
langchain-core==1.1.1

# Data Processing
pandas==2.2.2

# Utilities
python-dotenv==1.1.1
aiohttp==3.13.2
